fix: preprocessing and training changes
what: training erros and preprocesing fixes.
why: we were performingthe training on the whole dataset instead the
training set. Also, mean and std were miscalculated.
how: now the training and testing set are created as single files in a
separated folder. Then, the mean and std are calculated. This information
is fed into the training process via arguments, so launch.json was
updated.